+++ packages/sql/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./sqlPersistence";
+export * from "./sqlWorldStore";